Suite processor extracted
